+++ src/components/Icon/Icon.tsx
@@ -1,274 +1,7 @@
 import React from "react";
 import { SVGAttributes } from "react";
 import { Theme } from "../../styles/theme";
-
-const iconList = {
-  add: (
-    <path
-      d="M4.5 12.6H20.5M12.5 4.59998V20.6"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-    />
-  ),
-  alarm: (
-    <>
-      <path
-        d="M20.5 18.6H4.5L5.64286 15.1384V10.4571C5.64286 6.67002 8.7129 3.59998 12.5 3.59998C16.2871 3.59998 19.3571 6.67002 19.3571 10.4571V15.1384L20.5 18.6Z"
-        stroke="currentColor"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M9.5 18.6V18.6C9.5 20.2568 10.8431 21.6 12.5 21.6V21.6C14.1569 21.6 15.5 20.2568 15.5 18.6V18.6"
-        stroke="currentColor"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </>
-  ),
-  arrowLeft: (
-    <path
-      d="M17.5 3.19995L8.5 12.2L17.5 21.2"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  ),
-  arrowRight: (
-    <path
-      d="M8.5 3.19995L17.5 12.2L8.5 21.2"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  ),
-  calendar: (
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M8.35 2.65002H7.2V3.90002H8.35L8.35 2.65002ZM6.2 2.65002V3.90002H4C2.89543 3.90002 2 4.79545 2 5.90002V7.90002V8.90002V20.9C2 22.0046 2.89543 22.9 4 22.9H21C22.1046 22.9 23 22.0046 23 20.9V8.90002V7.90002V5.90002C23 4.79545 22.1046 3.90002 21 3.90002H18.8V2.65002C18.8 2.09774 18.3523 1.65002 17.8 1.65002H16.65C16.0977 1.65002 15.65 2.09774 15.65 2.65002V3.90002H9.35V2.65002C9.35 2.09774 8.90229 1.65002 8.35 1.65002H7.2C6.64772 1.65002 6.2 2.09774 6.2 2.65002ZM16.65 3.90002H17.8V2.65002H16.65V3.90002ZM15.65 4.90002H9.35H8.35H7.2H6.2H4C3.44772 4.90002 3 5.34774 3 5.90002V7.90002H22V5.90002C22 5.34774 21.5523 4.90002 21 4.90002H18.8H17.8H16.65H15.65ZM3 8.90002H22V20.9C22 21.4523 21.5523 21.9 21 21.9H4C3.44772 21.9 3 21.4523 3 20.9V8.90002ZM6.2 11.9H8.71999V14.2334H6.2V11.9ZM11.24 11.9H13.76V14.2334H11.24V11.9ZM18.8 11.9H16.28V14.2334H18.8V11.9ZM6.2 16.5667H8.71999V18.9H6.2V16.5667ZM13.76 16.5667H11.24V18.9H13.76V16.5667ZM16.28 16.5667H18.8V18.9H16.28V16.5667Z"
-      fill="currentColor"
-    />
-  ),
-  cancel: (
-    <path
-      d="M19.5 5L5.5 19M5.5 5L19.5 19"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  ),
-  cancleWithCircle: (
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M12.5 22C18.0228 22 22.5 17.5228 22.5 12C22.5 6.47715 18.0228 2 12.5 2C6.97715 2 2.5 6.47715 2.5 12C2.5 17.5228 6.97715 22 12.5 22ZM17.2211 16.7211C16.8491 17.093 16.2461 17.093 15.8742 16.7211L12.5 13.3469L9.12582 16.7211C8.75389 17.093 8.15087 17.093 7.77895 16.7211C7.40702 16.3491 7.40702 15.7461 7.77895 15.3742L11.1531 12L7.77895 8.62582C7.40702 8.25389 7.40702 7.65087 7.77895 7.27895C8.15087 6.90702 8.75389 6.90702 9.12582 7.27895L12.5 10.6531L15.8742 7.27895C16.2461 6.90702 16.8491 6.90702 17.2211 7.27895C17.593 7.65087 17.593 8.25389 17.2211 8.62582L13.8469 12L17.2211 15.3742C17.593 15.7461 17.593 16.3491 17.2211 16.7211Z"
-      fill="#9E9E9E"
-    />
-  ),
-  check: (
-    <path
-      d="M4.5 12.3L10 18.8L20 6.79999"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  ),
-  checkInCircle: (
-    <>
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M19.5 12.2C19.5 16.0659 16.366 19.2 12.5 19.2C8.63401 19.2 5.5 16.0659 5.5 12.2C5.5 8.33396 8.63401 5.19995 12.5 5.19995C14.0503 5.19995 15.4829 5.70394 16.6429 6.55702L18.0699 5.13007C16.5377 3.9213 14.6031 3.19995 12.5 3.19995C7.52944 3.19995 3.5 7.22939 3.5 12.2C3.5 17.1705 7.52944 21.2 12.5 21.2C17.4706 21.2 21.5 17.1705 21.5 12.2C21.5 11.1514 21.3207 10.1447 20.991 9.20891L19.3667 10.8333C19.4541 11.2753 19.5 11.7323 19.5 12.2Z"
-        fill="currentColor"
-      />
-      <path
-        d="M11.1063 11.392L9.09261 12.2068L11.7929 14.9071C12.1834 15.2976 12.8166 15.2976 13.2071 14.9071L20.5389 7.5753L15.7882 9.49756L12.5 12.7857L11.1063 11.392Z"
-        fill="currentColor"
-      />
-      <path
-        d="M19.4191 5.86662L21.9969 4.33187C21.6149 4.11288 21.1192 4.16654 20.7929 4.49284L19.4191 5.86662Z"
-        fill="currentColor"
-      />
-      <path
-        d="M22.2071 5.90706C22.5976 5.51653 22.5976 4.88337 22.2071 4.49284C22.1429 4.42862 22.0721 4.37497 21.9969 4.33187L19.4191 5.86662L15.7882 9.49756L20.5389 7.5753L22.2071 5.90706Z"
-        fill="currentColor"
-      />
-      <path
-        d="M9.09261 12.2068L11.1063 11.392L10.7451 11.0309L8.97223 12.0864L9.09261 12.2068Z"
-        fill="currentColor"
-      />
-      <path
-        d="M10.2071 10.4928C9.81658 10.1023 9.18342 10.1023 8.79289 10.4928C8.40237 10.8834 8.40237 11.5165 8.79289 11.9071L8.97223 12.0864L10.7451 11.0309L10.2071 10.4928Z"
-        fill="currentColor"
-      />
-    </>
-  ),
-  checkWithCircle: (
-    <>
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M12.5 22.8C18.0228 22.8 22.5 18.3228 22.5 12.8C22.5 7.27714 18.0228 2.79999 12.5 2.79999C6.97715 2.79999 2.5 7.27714 2.5 12.8C2.5 18.3228 6.97715 22.8 12.5 22.8ZM18.2481 9.4636C18.6146 9.05045 18.5768 8.41842 18.1636 8.05191C17.7505 7.68541 17.1184 7.72322 16.7519 8.13637L10.3975 15.2995L7.24302 11.7974C6.87339 11.387 6.24109 11.354 5.83073 11.7236C5.42037 12.0933 5.38735 12.7256 5.75698 13.1359L9.66021 17.4693C9.85072 17.6808 10.1224 17.8011 10.407 17.8C10.6917 17.7989 10.9624 17.6766 11.1513 17.4636L18.2481 9.4636Z"
-        fill="currentColor"
-      />
-    </>
-  ),
-  delete: (
-    <path
-      d="M4.5 6.29998H6.5M20.5 6.29998H18.5M6.5 6.29998V19.6C6.5 20.7045 7.39543 21.6 8.5 21.6H16.5C17.6046 21.6 18.5 20.7045 18.5 19.6V6.29998M6.5 6.29998H10.5M18.5 6.29998H14.5M10.5 9.89998V18M14.5 18V9.89998M10.5 6.29998V4.59998C10.5 4.04769 10.9477 3.59998 11.5 3.59998H13.5C14.0523 3.59998 14.5 4.04769 14.5 4.59998V6.29998M10.5 6.29998H14.5"
-      stroke="currentColor"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  ),
-  down: (
-    <path
-      d="M6.79289 9.29289C7.18342 8.90237 7.81658 8.90237 8.20711 9.29289L12.5 13.5858L16.7929 9.29289C17.1834 8.90237 17.8166 8.90237 18.2071 9.29289C18.5976 9.68342 18.5976 10.3166 18.2071 10.7071L13.2071 15.7071C12.8166 16.0976 12.1834 16.0976 11.7929 15.7071L6.79289 10.7071C6.40237 10.3166 6.40237 9.68342 6.79289 9.29289Z"
-      fill="currentColor"
-    />
-  ),
-  downWithCircle: (
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M12.5 22C18.0228 22 22.5 17.5228 22.5 12C22.5 6.47715 18.0228 2 12.5 2C6.97715 2 2.5 6.47715 2.5 12C2.5 17.5228 6.97715 22 12.5 22ZM8.20711 9.29289C7.81658 8.90237 7.18342 8.90237 6.79289 9.29289C6.40237 9.68342 6.40237 10.3166 6.79289 10.7071L11.7929 15.7071C12.1834 16.0976 12.8166 16.0976 13.2071 15.7071L18.2071 10.7071C18.5976 10.3166 18.5976 9.68342 18.2071 9.29289C17.8166 8.90237 17.1834 8.90237 16.7929 9.29289L12.5 13.5858L8.20711 9.29289Z"
-      fill="currentColor"
-    />
-  ),
-  exclamat: (
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M13.5909 7.07272C13.5909 6.36981 13.0109 5.79999 12.2954 5.79999C11.58 5.79999 11 6.36981 11 7.07272V14.7091C11 15.412 11.58 15.9818 12.2954 15.9818C13.0109 15.9818 13.5909 15.412 13.5909 14.7091V7.07272ZM12.2954 19.8001C13.0109 19.8001 13.5909 19.2302 13.5909 18.5273C13.5909 17.8244 13.0109 17.2546 12.2954 17.2546C11.58 17.2546 11 17.8244 11 18.5273C11 19.2302 11.58 19.8001 12.2954 19.8001Z"
-      fill="currentColor"
-    />
-  ),
-  filter: (
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M15.3158 3.17365C15.5919 3.17365 15.8158 3.3975 15.8158 3.67365V5.8003V7.92695C15.8158 8.2031 15.5919 8.42695 15.3158 8.42695C15.0396 8.42695 14.8158 8.2031 14.8158 7.92695V6.3003H3.97363C3.69749 6.3003 3.47363 6.07644 3.47363 5.8003C3.47363 5.52416 3.69749 5.3003 3.97363 5.3003H14.8158V3.67365C14.8158 3.3975 15.0396 3.17365 15.3158 3.17365ZM17.7061 5.8003C17.7061 5.52416 17.93 5.3003 18.2061 5.3003H20.9869C21.263 5.3003 21.4869 5.52416 21.4869 5.8003C21.4869 6.07644 21.263 6.3003 20.9869 6.3003H18.2061C17.93 6.3003 17.7061 6.07644 17.7061 5.8003ZM7.75435 9.12838C8.03049 9.12838 8.25435 9.35224 8.25435 9.62838V12.1804V14.7324C8.25435 15.0085 8.03049 15.2324 7.75435 15.2324C7.47821 15.2324 7.25435 15.0085 7.25435 14.7324V12.6804H3.97363C3.69749 12.6804 3.47363 12.4565 3.47363 12.1804C3.47363 11.9042 3.69749 11.6804 3.97363 11.6804H7.25435V9.62838C7.25435 9.35224 7.47821 9.12838 7.75435 9.12838ZM10.1447 12.1804C10.1447 11.9042 10.3686 11.6804 10.6447 11.6804H20.9869C21.263 11.6804 21.4869 11.9042 21.4869 12.1804C21.4869 12.4565 21.263 12.6804 20.9869 12.6804H10.6447C10.3686 12.6804 10.1447 12.4565 10.1447 12.1804ZM12.9802 16.4336C12.9802 16.1574 12.7564 15.9336 12.4802 15.9336C12.2041 15.9336 11.9802 16.1574 11.9802 16.4336V18.0602H3.97363C3.69749 18.0602 3.47363 18.2841 3.47363 18.5602C3.47363 18.8364 3.69749 19.0602 3.97363 19.0602H11.9802V20.6869C11.9802 20.963 12.2041 21.1869 12.4802 21.1869C12.7564 21.1869 12.9802 20.963 12.9802 20.6869V18.5602V16.4336ZM15.3706 18.0602C15.0945 18.0602 14.8706 18.2841 14.8706 18.5602C14.8706 18.8364 15.0945 19.0602 15.3706 19.0602H20.9869C21.263 19.0602 21.4869 18.8364 21.4869 18.5602C21.4869 18.2841 21.263 18.0602 20.9869 18.0602H15.3706Z"
-      fill="currentColor"
-    />
-  ),
-  flag: (
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M7 4.40002C7 4.12388 6.77614 3.90002 6.5 3.90002C6.22386 3.90002 6 4.12388 6 4.40002V6.65002V13.65V20.4C6 20.6762 6.22386 20.9 6.5 20.9C6.77614 20.9 7 20.6762 7 20.4V14.65H19C19.961 14.65 20.3688 13.4266 19.6 12.85L16.8333 10.775L16 10.15L16.8333 9.52502L19.6 7.45002C20.3688 6.87342 19.961 5.65002 19 5.65002H7V4.40002ZM7 6.65002V13.65H19L15.4 10.95C15.1482 10.7612 15 10.4648 15 10.15C15 9.83527 15.1482 9.53888 15.4 9.35002L19 6.65002H7Z"
-      fill="currentColor"
-    />
-  ),
-  github: (
-    <path
-      d="M12.5011 2C6.97548 1.99771 2.5 6.59261 2.5 12.2639C2.5 16.7487 5.29969 20.5609 9.19869 21.961C9.72377 22.0964 9.64334 21.7131 9.64334 21.4515V19.6727C6.61127 20.0376 6.48838 17.9766 6.28505 17.6323C5.87392 16.9116 4.90197 16.728 5.19244 16.3837C5.88286 16.0188 6.58669 16.4756 7.40225 17.7126C7.99212 18.6101 9.14283 18.4586 9.72601 18.3094C9.85337 17.77 10.126 17.288 10.5013 16.9139C7.35979 16.3356 6.05044 14.3663 6.05044 12.0252C6.05044 10.8891 6.41465 9.84485 7.12965 9.00253C6.67384 7.61396 7.1721 6.42506 7.23914 6.24834C8.53731 6.12899 9.88688 7.20312 9.9919 7.28804C10.7292 7.08377 11.5716 6.9759 12.5145 6.9759C13.4619 6.9759 14.3065 7.08836 15.0506 7.29493C15.303 7.09754 16.5543 6.17489 17.7609 6.28735C17.8257 6.46408 18.3128 7.62543 17.8838 8.99564C18.6077 9.84026 18.9764 10.8937 18.9764 12.0321C18.9764 14.3778 17.6581 16.3493 14.5076 16.9185C14.7774 17.1911 14.9917 17.5162 15.1379 17.8748C15.284 18.2334 15.3592 18.6183 15.3589 19.0071V21.5892C15.3768 21.7957 15.3589 22 15.6941 22C19.6512 20.6298 22.5 16.79 22.5 12.2662C22.5 6.59261 18.0223 2 12.5011 2V2Z"
-      fill="currentColor"
-    />
-  ),
-  google: (
-    <path
-      d="M22.3256 10.1798H13.1934V14.0495H18.4103C18.1854 15.3003 17.5029 16.3583 16.4742 17.0671C15.6047 17.6664 14.4951 18.0208 13.1908 18.0208C10.6658 18.0208 8.52993 16.2619 7.7666 13.8984C7.5745 13.299 7.46328 12.658 7.46328 11.9987C7.46328 11.3394 7.5745 10.6984 7.7666 10.099C8.53246 7.73811 10.6683 5.97915 13.1934 5.97915C14.6164 5.97915 15.8928 6.48469 16.8988 7.47492L19.6792 4.60586C17.9983 2.99023 15.8069 2 13.1934 2C9.40448 2 6.12618 4.24104 4.53126 7.50879C3.87409 8.85863 3.5 10.3857 3.5 12.0013C3.5 13.6169 3.87409 15.1414 4.53126 16.4912C6.12618 19.759 9.40448 22 13.1934 22C15.812 22 18.0059 21.1036 19.6084 19.5765C21.4409 17.8358 22.5 15.2717 22.5 12.2254C22.5 11.5166 22.4393 10.8365 22.3256 10.1798Z"
-      fill="currentColor"
-    />
-  ),
-  hamburger: (
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M3 4.65771C3 4.10543 3.44772 3.65771 4 3.65771H21C21.5523 3.65771 22 4.10543 22 4.65771C22 5.21 21.5523 5.65771 21 5.65771H4C3.44772 5.65771 3 5.21 3 4.65771ZM3 12.6519C3 12.0996 3.44772 11.6519 4 11.6519H21C21.5523 11.6519 22 12.0996 22 12.6519C22 13.2042 21.5523 13.6519 21 13.6519H4C3.44772 13.6519 3 13.2042 3 12.6519ZM4 19.6461C3.44772 19.6461 3 20.0938 3 20.6461C3 21.1984 3.44772 21.6461 4 21.6461H21C21.5523 21.6461 22 21.1984 22 20.6461C22 20.0938 21.5523 19.6461 21 19.6461H4Z"
-      fill="currentColor"
-    />
-  ),
-  kakaotalk: (
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M12.5 18.5517C18.0228 18.5517 22.5 14.8465 22.5 10.2758C22.5 5.70522 18.0228 2 12.5 2C6.97715 2 2.5 5.70522 2.5 10.2758C2.5 13.1637 4.28734 15.7061 6.99648 17.1867L6.01769 21.2368C5.95301 21.5045 6.25293 21.7106 6.47962 21.5543L10.9725 18.4557C11.4705 18.5189 11.9806 18.5517 12.5 18.5517Z"
-      fill="currentColor"
-    />
-  ),
-  location: (
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M12.2757 3.90002C8.80973 3.90002 6 6.70975 6 10.1757C6 11.5861 6.47506 12.9553 7.34857 14.0626L12.5 20.5927L17.6514 14.0626C18.5249 12.9553 19 11.5861 19 10.1757C19 6.70975 16.1903 3.90002 12.7243 3.90002H12.2757ZM5 10.1757C5 6.15747 8.25744 2.90002 12.2757 2.90002H12.7243C16.7426 2.90002 20 6.15747 20 10.1757C20 11.8108 19.4492 13.3982 18.4365 14.682L12.8926 21.7097C12.7977 21.8299 12.6531 21.9 12.5 21.9C12.3469 21.9 12.2023 21.8299 12.1074 21.7097L6.56345 14.682C5.55076 13.3982 5 11.8108 5 10.1757ZM12.5 12.4C13.6046 12.4 14.5 11.5046 14.5 10.4C14.5 9.29545 13.6046 8.40002 12.5 8.40002C11.3954 8.40002 10.5 9.29545 10.5 10.4C10.5 11.5046 11.3954 12.4 12.5 12.4ZM12.5 13.4C14.1569 13.4 15.5 12.0569 15.5 10.4C15.5 8.74317 14.1569 7.40002 12.5 7.40002C10.8431 7.40002 9.5 8.74317 9.5 10.4C9.5 12.0569 10.8431 13.4 12.5 13.4Z"
-      fill="currentColor"
-    />
-  ),
-  question: (
-    <path
-      d="M11.3185 19.3247C11.3185 18.9239 11.4396 18.5899 11.6816 18.3226C11.9379 18.0554 12.2512 17.9218 12.6214 17.9218C13.0201 17.9218 13.3334 18.0554 13.5612 18.3226C13.8033 18.5899 13.9243 18.9239 13.9243 19.3247C13.9243 19.7107 13.8033 20.0373 13.5612 20.3046C13.3334 20.5718 13.0201 20.7054 12.6214 20.7054C12.2512 20.7054 11.9379 20.5718 11.6816 20.3046C11.4396 20.0373 11.3185 19.7107 11.3185 19.3247ZM8.51905 7.99359C8.20007 7.64519 8.1771 7.11153 8.50485 6.77137C8.98227 6.27587 9.49299 5.87295 10.037 5.5626C10.8059 5.11722 11.6888 4.89453 12.6855 4.89453C13.4829 4.89453 14.202 5.07268 14.8427 5.42898C15.4977 5.77044 16.0175 6.26035 16.4019 6.89873C16.8006 7.52225 17 8.26455 17 9.12561C17 9.64522 16.9003 10.1351 16.701 10.5954C16.5016 11.0556 16.2168 11.4713 15.8466 11.8424C15.4906 12.2136 15.0777 12.5476 14.6078 12.8445C14.3372 13.0078 14.1237 13.1711 13.967 13.3344C13.8104 13.4977 13.7036 13.6833 13.6466 13.8911C13.5897 14.099 13.547 14.3662 13.5185 14.6928L13.5037 14.9498C13.476 15.4306 13.0781 15.8063 12.5965 15.8063C12.1267 15.8063 11.7344 15.4481 11.6916 14.9802L11.6389 14.4033C11.582 13.7946 11.6674 13.2825 11.8952 12.8668C12.123 12.4511 12.4648 12.0799 12.9204 11.7533C13.3476 11.4564 13.7036 11.1743 13.9884 10.9071C14.2874 10.6399 14.5081 10.3652 14.6505 10.0832C14.8071 9.8011 14.8855 9.48191 14.8855 9.12561C14.8855 8.69508 14.7858 8.31651 14.5864 7.9899C14.4013 7.66329 14.1379 7.40349 13.7962 7.21049C13.4544 7.01749 13.0628 6.921 12.6214 6.921C11.8952 6.921 11.2758 7.09172 10.7632 7.43318C10.5289 7.58244 10.3006 7.7627 10.0783 7.97397C9.63869 8.39161 8.9285 8.44081 8.51905 7.99359Z"
-      fill="currentColor"
-    />
-  ),
-  questionWithCircle: (
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M12.5 22.8C18.0228 22.8 22.5 18.3228 22.5 12.8C22.5 7.27714 18.0228 2.79999 12.5 2.79999C6.97715 2.79999 2.5 7.27714 2.5 12.8C2.5 18.3228 6.97715 22.8 12.5 22.8ZM11.7908 17.6901C11.5764 17.9267 11.4693 18.2225 11.4693 18.5775C11.4693 18.9192 11.5764 19.2084 11.7908 19.4451C12.0177 19.6817 12.2951 19.8 12.6229 19.8C12.976 19.8 13.2533 19.6817 13.4551 19.4451C13.6694 19.2084 13.7766 18.9192 13.7766 18.5775C13.7766 18.2225 13.6694 17.9267 13.4551 17.6901C13.2533 17.4535 12.976 17.3352 12.6229 17.3352C12.2951 17.3352 12.0177 17.4535 11.7908 17.6901ZM8.97784 7.46187C8.68763 7.76307 8.70798 8.2356 8.99042 8.5441C9.35297 8.9401 9.98182 8.89653 10.3711 8.52672C10.5679 8.33965 10.7701 8.18004 10.9775 8.04788C11.4314 7.74553 11.9799 7.59435 12.6229 7.59435C13.0138 7.59435 13.3605 7.6798 13.6631 7.85069C13.9657 8.02158 14.199 8.25163 14.3629 8.54083C14.5394 8.83003 14.6277 9.16525 14.6277 9.54647C14.6277 9.86196 14.5583 10.1446 14.4196 10.3944C14.2935 10.6441 14.0981 10.8873 13.8333 11.1239C13.5812 11.3606 13.266 11.6103 12.8877 11.8732C12.4842 12.1624 12.1816 12.4911 11.9799 12.8591C11.7782 13.2272 11.7025 13.6807 11.753 14.2197L11.7996 14.7305C11.8375 15.1448 12.1849 15.462 12.6009 15.462C13.0273 15.462 13.3797 15.1293 13.4042 14.7036L13.4173 14.476C13.4425 14.1868 13.4803 13.9502 13.5307 13.7662C13.5812 13.5821 13.6757 13.4178 13.8144 13.2732C13.9531 13.1286 14.1422 12.984 14.3818 12.8394C14.7979 12.5765 15.1635 12.2807 15.4787 11.9521C15.8065 11.6235 16.0587 11.2554 16.2352 10.8479C16.4117 10.4404 16.5 10.0066 16.5 9.54647C16.5 8.78403 16.3235 8.12675 15.9704 7.57464C15.63 7.00938 15.1698 6.57557 14.5898 6.27323C14.0225 5.95773 13.3857 5.79999 12.6797 5.79999C11.7971 5.79999 11.0154 5.99717 10.3345 6.39154C9.85281 6.66634 9.40058 7.02312 8.97784 7.46187Z"
-      fill="currentColor"
-    />
-  ),
-  scrapOff: (
-    <path
-      d="M19.5 3H5.5C4.94772 3 4.5 3.44772 4.5 4V20.5858C4.5 21.4767 5.57714 21.9229 6.2071 21.2929L11.7929 16.5C12.1834 16.1095 12.8166 16.1095 13.2071 16.5L18.7929 21.2929C19.4229 21.9229 20.5 21.4767 20.5 20.5858V4C20.5 3.44772 20.0523 3 19.5 3Z"
-      stroke="currentColor"
-      strokeWidth="2"
-    />
-  ),
-  scrapOn: (
-    <path
-      d="M19.5 3H5.5C4.94772 3 4.5 3.44772 4.5 4V20.5858C4.5 21.4767 5.57714 21.9229 6.2071 21.2929L11.7929 16.5C12.1834 16.1095 12.8166 16.1095 13.2071 16.5L18.7929 21.2929C19.4229 21.9229 20.5 21.4767 20.5 20.5858V4C20.5 3.44772 20.0523 3 19.5 3Z"
-      fill="currentColor"
-    />
-  ),
-  search: (
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M5.32117 11.541C5.32117 8.45568 7.83884 5.9444 10.9571 5.9444C14.0753 5.9444 16.5929 8.45568 16.5929 11.541C16.5929 14.6264 14.0753 17.1377 10.9571 17.1377C7.83884 17.1377 5.32117 14.6264 5.32117 11.541ZM10.9571 3.9444C6.74547 3.9444 3.32117 7.33994 3.32117 11.541C3.32117 15.7421 6.74547 19.1377 10.9571 19.1377C12.825 19.1377 14.538 18.4697 15.8661 17.3598L20.2059 21.6739C20.5976 22.0633 21.2307 22.0614 21.6201 21.6697C22.0094 21.278 22.0076 20.6449 21.6159 20.2555L17.2211 15.8867C18.0855 14.656 18.5929 13.1582 18.5929 11.541C18.5929 7.33994 15.1686 3.9444 10.9571 3.9444Z"
-      fill="currentColor"
-    />
-  ),
-  setting: (
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M13.7223 3.40003L11.2778 3.40002V4.19855C11.2778 5.10176 10.6852 5.81931 9.95079 6.11765C9.92092 6.12979 9.89116 6.14213 9.86151 6.15467C9.13096 6.46374 8.20405 6.37549 7.56516 5.7366L7.00032 5.17176L5.27183 6.90024L5.83663 7.46504C6.47554 8.10395 6.56378 9.03088 6.25469 9.76143C6.24213 9.79111 6.22978 9.82091 6.21763 9.85081C5.91928 10.5852 5.20173 11.1778 4.29852 11.1778H3.5L3.5 13.6223H4.29857C5.20177 13.6223 5.91931 14.2149 6.21766 14.9493C6.22979 14.9791 6.24213 15.0089 6.25468 15.0385C6.56376 15.7691 6.47551 16.696 5.83661 17.3349L5.2718 17.8997L7.00028 19.6282L7.5651 19.0634C8.204 18.4245 9.13092 18.3362 9.86147 18.6453C9.89113 18.6579 9.92091 18.6702 9.95079 18.6823C10.6852 18.9807 11.2778 19.6982 11.2778 20.6014V21.4H13.7223V20.6014C13.7223 19.6982 14.3148 18.9807 15.0492 18.6823C15.0791 18.6702 15.1089 18.6578 15.1385 18.6453C15.8691 18.3362 16.796 18.4244 17.4349 19.0633L17.9998 19.6282L19.7282 17.8997L19.1634 17.3348C18.5245 16.6959 18.4362 15.769 18.7453 15.0385C18.7578 15.0088 18.7702 14.9791 18.7823 14.9493C19.0806 14.2149 19.7982 13.6223 20.7014 13.6223H21.5V11.1778H20.7014C19.7982 11.1778 19.0807 10.5852 18.7823 9.85081C18.7702 9.82093 18.7578 9.79116 18.7453 9.76149C18.4362 9.03094 18.5244 8.10403 19.1633 7.46513L19.7282 6.90027L17.9997 5.17179L17.4349 5.73663C16.796 6.37554 15.869 6.46378 15.1385 6.1547C15.1089 6.14216 15.0791 6.12982 15.0492 6.11769C14.3148 5.81933 13.7223 5.10179 13.7223 4.1986V3.40003ZM11.2778 2.40002C10.7255 2.40002 10.2778 2.84774 10.2778 3.40002V4.19855C10.2778 4.64097 9.98432 5.02467 9.57443 5.19118C9.54012 5.20512 9.50593 5.21929 9.47187 5.2337C9.06427 5.40615 8.58522 5.34244 8.27226 5.02949L7.70742 4.46465C7.3169 4.07412 6.68373 4.07412 6.29321 4.46465L4.56473 6.19313C4.1742 6.58366 4.1742 7.21682 4.56472 7.60735L5.12953 8.17215C5.44248 8.48511 5.50619 8.96417 5.33373 9.37177C5.3193 9.40588 5.30511 9.4401 5.29116 9.47445C5.12465 9.88434 4.74095 10.1778 4.29852 10.1778L3.5 10.1778C2.94771 10.1778 2.5 10.6255 2.5 11.1778V13.6223C2.5 14.1746 2.94772 14.6223 3.5 14.6223H4.29857C4.74099 14.6223 5.12468 14.9158 5.2912 15.3256C5.30513 15.3599 5.3193 15.3941 5.33371 15.4282C5.50616 15.8358 5.44246 16.3148 5.1295 16.6278L4.56469 17.1926C4.17417 17.5831 4.17416 18.2163 4.56469 18.6068L6.29317 20.3353C6.6837 20.7258 7.31686 20.7258 7.70739 20.3353L8.27221 19.7705C8.58516 19.4575 9.06422 19.3938 9.47183 19.5663C9.5059 19.5807 9.5401 19.5949 9.57443 19.6088C9.98432 19.7753 10.2778 20.159 10.2778 20.6014V21.4C10.2778 21.9523 10.7255 22.4 11.2778 22.4H13.7223C14.2745 22.4 14.7223 21.9523 14.7223 21.4V20.6014C14.7223 20.159 15.0157 19.7753 15.4256 19.6088C15.4599 19.5948 15.4941 19.5806 15.5282 19.5662C15.9358 19.3938 16.4149 19.4575 16.7278 19.7704L17.2926 20.3353C17.6832 20.7258 18.3163 20.7258 18.7069 20.3353L20.4353 18.6068C20.8259 18.2163 20.8259 17.5831 20.4353 17.1926L19.8705 16.6277C19.5575 16.3148 19.4938 15.8357 19.6663 15.4281C19.6807 15.3941 19.6948 15.3599 19.7087 15.3256C19.8753 14.9158 20.259 14.6223 20.7014 14.6223H21.5C22.0523 14.6223 22.5 14.1746 22.5 13.6223V11.1778C22.5 10.6255 22.0523 10.1778 21.5 10.1778H20.7014C20.259 10.1778 19.8753 9.88434 19.7088 9.47445C19.6948 9.44013 19.6807 9.40593 19.6662 9.37185C19.4938 8.96424 19.5575 8.48519 19.8705 8.17223L20.4353 7.60738C20.8258 7.21685 20.8258 6.58369 20.4353 6.19317L18.7068 4.46468C18.3163 4.07416 17.6831 4.07416 17.2926 4.46468L16.7278 5.02953C16.4148 5.34248 15.9358 5.40619 15.5281 5.23374C15.4941 5.21933 15.4599 5.20516 15.4256 5.19122C15.0157 5.02471 14.7223 4.64101 14.7223 4.1986V3.40003C14.7223 2.84774 14.2745 2.40002 13.7223 2.40002H11.2778ZM12.4999 15.4C14.1568 15.4 15.4999 14.0569 15.4999 12.4C15.4999 10.7432 14.1568 9.40004 12.4999 9.40004C10.8431 9.40004 9.49995 10.7432 9.49995 12.4C9.49995 14.0569 10.8431 15.4 12.4999 15.4ZM12.4999 16.4C14.7091 16.4 16.4999 14.6092 16.4999 12.4C16.4999 10.1909 14.7091 8.40004 12.4999 8.40004C10.2908 8.40004 8.49995 10.1909 8.49995 12.4C8.49995 14.6092 10.2908 16.4 12.4999 16.4Z"
-      fill="currentColor"
-    />
-  ),
-  share: (
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M18.587 5.40002C18.587 6.51201 17.6806 7.40002 16.5791 7.40002C16.1242 7.40002 15.7027 7.2486 15.3644 6.99251C15.345 7.01316 15.3235 7.03243 15.3 7.05002L9.31937 11.5355C9.44518 11.7966 9.51582 12.0896 9.51582 12.4C9.51582 12.7107 9.44508 13.0039 9.3191 13.2651C9.31944 13.2654 9.31978 13.2656 9.32012 13.2659L15.0818 18.0673C15.4507 17.6569 15.9867 17.4 16.5791 17.4C17.6806 17.4 18.587 18.288 18.587 19.4C18.587 20.512 17.6806 21.4 16.5791 21.4C15.4776 21.4 14.5712 20.512 14.5712 19.4C14.5712 19.2559 14.5864 19.1155 14.6153 18.9803L8.67994 14.0341C8.67774 14.0323 8.67556 14.0305 8.6734 14.0286C8.34388 14.2628 7.94097 14.4 7.50791 14.4C6.40642 14.4 5.5 13.512 5.5 12.4C5.5 11.288 6.40642 10.4 7.50791 10.4C7.94096 10.4 8.34386 10.5373 8.67339 10.7714C8.68196 10.7641 8.69084 10.7569 8.70003 10.75L14.7 6.25002C14.7151 6.23873 14.7306 6.22841 14.7464 6.21906C14.6339 5.96961 14.5712 5.69257 14.5712 5.40002C14.5712 4.28804 15.4776 3.40002 16.5791 3.40002C17.6806 3.40002 18.587 4.28804 18.587 5.40002ZM16.5791 6.40002C17.1357 6.40002 17.587 5.95231 17.587 5.40002C17.587 4.84774 17.1357 4.40002 16.5791 4.40002C16.0224 4.40002 15.5712 4.84774 15.5712 5.40002C15.5712 5.95231 16.0224 6.40002 16.5791 6.40002ZM8.51582 12.4C8.51582 12.9523 8.06457 13.4 7.50791 13.4C6.95126 13.4 6.5 12.9523 6.5 12.4C6.5 11.8477 6.95126 11.4 7.50791 11.4C8.06457 11.4 8.51582 11.8477 8.51582 12.4ZM17.587 19.4C17.587 19.9523 17.1357 20.4 16.5791 20.4C16.0224 20.4 15.5712 19.9523 15.5712 19.4C15.5712 18.8477 16.0224 18.4 16.5791 18.4C17.1357 18.4 17.587 18.8477 17.587 19.4Z"
-      fill="currentColor"
-    />
-  ),
-  tag: (
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M19.9353 5.63528C19.9353 5.35914 19.7114 5.13528 19.4353 5.13528L12.9353 5.13528C12.8027 5.13528 12.6755 5.18796 12.5817 5.28173L4.13175 13.7317C3.93649 13.927 3.93649 14.2436 4.13175 14.4388L10.6317 20.9388C10.827 21.1341 11.1436 21.1341 11.3389 20.9388L19.7889 12.4888C19.8826 12.3951 19.9353 12.2679 19.9353 12.1353L19.9353 5.63528ZM5.19241 14.0853L13.1424 6.13528L18.9353 6.13528L18.9353 11.9282L10.9853 19.8782L5.19241 14.0853ZM14.8853 10.1853C15.2443 10.5443 15.8263 10.5443 16.1853 10.1853C16.5443 9.82635 16.5443 9.24432 16.1853 8.88533C15.8263 8.52635 15.2443 8.52635 14.8853 8.88533C14.5263 9.24432 14.5263 9.82635 14.8853 10.1853Z"
-      fill="currentColor"
-    />
-  ),
-  write: (
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M17.393 3.20065C17.5883 3.00538 17.9049 3.00538 18.1001 3.20065L21.8993 6.99987C22.0946 7.19513 22.0946 7.51171 21.8993 7.70697L8.60209 21.0043C8.5323 21.0741 8.4434 21.1217 8.3466 21.141L3.59758 22.0908C3.43365 22.1236 3.26418 22.0723 3.14597 21.9541C3.02776 21.8359 2.97644 21.6664 3.00923 21.5025L3.95903 16.7534C3.97839 16.6567 4.02597 16.5678 4.09577 16.498L17.393 3.20065ZM17.7466 4.26131L15.2787 6.72924L18.3708 9.82135L20.8387 7.35342L17.7466 4.26131ZM4.90992 17.098L14.5715 7.43635L17.6637 10.5285L8.00204 20.1901L4.1369 20.9632L4.90992 17.098ZM12 20.5999C11.7238 20.5999 11.5 20.8238 11.5 21.0999C11.5 21.3761 11.7238 21.5999 12 21.5999H21C21.2761 21.5999 21.5 21.3761 21.5 21.0999C21.5 20.8238 21.2761 20.5999 21 20.5999H12Z"
-      fill="currentColor"
-    />
-  ),
-};
+import iconList from "./iconList";
 
 export type IconName = keyof typeof iconList;
 export interface IconProps extends SVGAttributes<SVGElement> {
